requirements.txt
tflite-runtime==2.14.0
streamlit==1.35.0
scikit-learn==1.4.2
pandas==2.2.2
tqdm==4.66.2
opencv-python==4.9.0.80
Pillow==10.3.0
lime==0.2.0.1
matplotlib==3.8.4
pytest==8.2.1
flake8==7.0.0
black==24.4.2 